Change the bit detection to halfwave scanning
The SBC scanning method works OK for the original 300 baud
Kansas City Standard protocol. But it tends to fail for the
protocols with higher bitrates, that eventually encode a
ZERO bit in just a single cycle of 1/2 the base frequency.
Decoding the sign-bit-changes into halfwave patterns first
and deriving bits, then bytes from that is more reliable.