Improve the bit pattern detection
Apparently some interfaces, like for example the CASIO CE-122,
are not reliable on the number of waves they produce per bit.
If for example the protocol specifies 4 cycles of 2000Hz for a
ZERO bit and 8 cycles of 4000HZ for a ONE, then the interface
may sometimes emit 5 2000Hz cycles for a ZERO and only 7 2000Hz
cycles for the following ONE.

This commit changes the detection logic by handling startbits
separately and synchronizing the half-wave-input-buffer on
where they actually began. From there on the following data
bits are sampled by just looking at the frequency that was
detected in the middle of the bit's time frame. This technique
seems to be working for all the interfaces in my collection.