Add test cases to guard against regression
As we add more protocols we need to safeguard against regression
on previous protocols. The idea here is to have an audio recording
from an original calculator that can be decoded into text. Decode
that and check that it produces to expected text output. Next take
the text representation of that program or data, convert it into
audio and back to check that it still matches the expected output.